Fix Bugs

+++ src/main/java/app/mobile/BK_Kindergarten/children/ChildrenRepository.java
@@ -14,5 +14,5 @@ public interface ChildrenRepository extends JpaRepository<Children, Long> {
     List<Children> findAllByName(String name);
 
     @Query("SELECT c FROM Children c WHERE c.parent_id = :parent_id")
-    List<Children> findAllByparentId(Long parent_id);
+    List<Children> findAllByparentId(Integer parent_id);
 }
